+++ src/ValueObject.php
@@ -21,6 +21,8 @@ use ZeroBoiler\ValueObjects\Contracts\ValueObject as ValueObjectContract;
  *
  * All value objects implement {@see ValueObjectContract} which serves
  * as the cross-package type for accepting arbitrary value objects.
+ *
+ * @since 1.0.0
  */
 abstract class ValueObject implements ValueObjectInterface
 {